src/car-detector-dataset/pyproject.toml
[project]
name = "car-detector-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy==1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "tensorflow-macos==2.16.*",
    "tqdm>=4.67.1",
]
